refactor: useTabMenu 로직 분리
